# docs/reference/openapi.json
{
  "components": {
    "schemas": {
...
      "ConformanceIssue": {
        "properties": {
          "level": {
            "enum": [
              "warning",
              "nonconformant"
            ],
            "title": "Level",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "path": {
            "title": "Path",
            "type": "string"
          },
          "rule": {
            "title": "Rule",
            "type": "string"
          }
        },
        "required": [
          "path",
          "level",
          "rule",
          "message"
        ],
        "title": "ConformanceIssue",
        "type": "object"
      },
      "ConformanceReport": {
        "properties": {
          "effective_schema_hash": {
            "title": "Effective Schema Hash",
            "type": "string"
          },
          "issues": {
            "items": {
              "$ref": "#/components/schemas/ConformanceIssue"
            },
            "title": "Issues",
            "type": "array"
          },
          "status": {
            "enum": [
              "ok",
              "warning",
              "nonconformant",
              "waived"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status",
          "issues",
          "effective_schema_hash"
        ],
        "title": "ConformanceReport",
        "type": "object"
      },
...
      "DefaultRelationOut": {
        "properties": {
          "relation": {
            "title": "Relation",
            "type": "string"
          }
        },
        "required": [
          "relation"
        ],
        "title": "DefaultRelationOut",
        "type": "object"
      },
      "DetectKeysBody": {
        "description": "Body of ``POST /keys/detect`` (a read: it only looks keys up).",
        "properties": {
          "linked_to": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Linked To"
          },
          "scope": {
            "title": "Scope",
            "type": "string"
          },
          "text": {
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "scope",
          "text"
        ],
        "title": "DetectKeysBody",
        "type": "object"
      },
...
      "EnumValue": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "crosswalk": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Crosswalk"
          },
          "label": {
            "title": "Label",
            "type": "string"
          }
        },
        "required": [
          "code",
          "label"
        ],
        "title": "EnumValue",
        "type": "object"
      },
      "ErrorBody": {
        "description": "The JSON body of every error response.",
        "properties": {
          "error": {
            "title": "Error",
            "type": "string"
          },
          "issues": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Issues"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "position": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Position"
          },
          "results": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Results"
          }
        },
        "required": [
          "error",
          "message"
        ],
        "title": "ErrorBody",
        "type": "object"
      },
...
      "FieldMeta": {
        "properties": {
          "description": {
            "title": "Description",
            "type": "string"
          },
          "enforcement": {
            "anyOf": [
              {
                "enum": [
                  "advisory",
                  "required",
                  "locked"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Enforcement"
          },
          "enum_values": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/EnumValue"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Enum Values"
          },
          "group": {
            "title": "Group",
            "type": "string"
          },
          "kind": {
            "enum": [
              "string",
              "text",
              "int",
              "decimal",
              "bool",
              "date",
              "datetime",
              "enum",
              "quantity",
              "ref",
              "json"
            ],
            "title": "Kind",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "layer": {
            "enum": [
              "core",
              "standard",
              "custom",
              "project",
              "enrichment",
              "source"
            ],
            "title": "Layer",
            "type": "string"
          },
          "order": {
            "title": "Order",
            "type": "integer"
          },
          "path": {
            "title": "Path",
            "type": "string"
          },
          "readonly": {
            "default": false,
            "title": "Readonly",
            "type": "boolean"
          },
          "required_in_states": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Required In States",
            "type": "array"
          },
          "unit": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unit"
          }
        },
        "required": [
          "path",
          "label",
          "description",
          "kind",
          "layer",
          "group",
          "order"
        ],
        "title": "FieldMeta",
        "type": "object"
      },
...
      "FormMetadata": {
        "properties": {
          "core_fields": {
            "items": {
              "$ref": "#/components/schemas/FieldMeta"
            },
            "title": "Core Fields",
            "type": "array"
          },
          "effective_schema_hash": {
            "title": "Effective Schema Hash",
            "type": "string"
          },
          "psets": {
            "items": {
              "$ref": "#/components/schemas/PsetGroupMeta"
            },
            "title": "Psets",
            "type": "array"
          },
          "record_type": {
            "title": "Record Type",
            "type": "string"
          }
        },
        "required": [
          "record_type",
          "effective_schema_hash",
          "core_fields",
          "psets"
        ],
        "title": "FormMetadata",
        "type": "object"
      },
      "GuardResult": {
        "properties": {
          "details": {
            "additionalProperties": true,
            "default": {},
            "title": "Details",
            "type": "object"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "passed": {
            "title": "Passed",
            "type": "boolean"
          }
        },
        "required": [
          "kind",
          "passed",
          "message"
        ],
        "title": "GuardResult",
        "type": "object"
      },
...
      "KeyChip": {
        "description": "A detected key and what it refers to.",
        "properties": {
          "end": {
            "title": "End",
            "type": "integer"
          },
          "key": {
            "title": "Key",
            "type": "string"
          },
          "link_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Link Status"
          },
          "pattern_id": {
            "title": "Pattern Id",
            "type": "string"
          },
          "record_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Record Id"
          },
          "record_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Record Type"
          },
          "start": {
            "title": "Start",
            "type": "integer"
          },
          "status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "voided": {
            "default": false,
            "title": "Voided",
            "type": "boolean"
          }
        },
        "required": [
          "start",
          "end",
          "key",
          "pattern_id",
          "record_id"
        ],
        "title": "KeyChip",
        "type": "object"
      },
...
      "PsetGroupMeta": {
        "properties": {
          "enforcement": {
            "anyOf": [
              {
                "enum": [
                  "advisory",
                  "required",
                  "locked"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Enforcement"
          },
          "fields": {
            "items": {
              "$ref": "#/components/schemas/FieldMeta"
            },
            "title": "Fields",
            "type": "array"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "layer": {
            "enum": [
              "core",
              "standard",
              "custom",
              "project",
              "enrichment",
              "source"
            ],
            "title": "Layer",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "package": {
            "title": "Package",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "string"
          }
        },
        "required": [
          "name",
          "label",
          "package",
          "version",
          "enforcement",
          "layer",
          "fields"
        ],
        "title": "PsetGroupMeta",
        "type": "object"
      },
...
      "RelationOut": {
        "description": "A relation of the link vocabulary (brief 7.1).",
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "inverse_code": {
            "title": "Inverse Code",
            "type": "string"
          },
          "inverse_label": {
            "title": "Inverse Label",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          }
        },
        "required": [
          "code",
          "label",
          "inverse_code",
          "inverse_label"
        ],
        "title": "RelationOut",
        "type": "object"
      },
...
      "TransitionOption": {
        "properties": {
          "allowed": {
            "title": "Allowed",
            "type": "boolean"
          },
          "from_state": {
            "title": "From State",
            "type": "string"
          },
          "guards": {
            "items": {
              "$ref": "#/components/schemas/GuardResult"
            },
            "title": "Guards",
            "type": "array"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "to_state": {
            "title": "To State",
            "type": "string"
          },
          "transition": {
            "title": "Transition",
            "type": "string"
          }
        },
        "required": [
          "transition",
          "label",
          "from_state",
          "to_state",
          "allowed",
          "guards"
        ],
        "title": "TransitionOption",
        "type": "object"
      },
...
      "WorkflowStatus": {
        "properties": {
          "entered_at": {
            "title": "Entered At",
            "type": "string"
          },
          "key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Key"
          },
          "options": {
            "items": {
              "$ref": "#/components/schemas/TransitionOption"
            },
            "title": "Options",
            "type": "array"
          },
          "record_id": {
            "title": "Record Id",
            "type": "string"
          },
          "state": {
            "title": "State",
            "type": "string"
          },
          "state_label": {
            "title": "State Label",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          },
          "workflow": {
            "title": "Workflow",
            "type": "string"
          },
          "workflow_version": {
            "title": "Workflow Version",
            "type": "integer"
          }
        },
        "required": [
          "record_id",
          "key",
          "workflow",
          "workflow_version",
          "state",
          "state_label",
          "entered_at",
          "version",
          "options"
        ],
        "title": "WorkflowStatus",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "description": "Static dev token (ADR-0005)",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Records, links, workflow, files and the change feed over HTTP. Phase 0 is a development server: identity is a static bearer token and every authenticated actor may do everything (ADR-0005). Errors are `{error, message, ...}` bodies; the `error` codes are stable.",
    "title": "Throughline API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/keys/detect": {
      "post": {
        "description": "Keys found in the text that fit a numbering pattern of the scope, resolved to records.",
        "operationId": "detect_keys",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DetectKeysBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/KeyChip"
                  },
                  "title": "Response Detect Keys",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Missing or unknown bearer token."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Refused by the authorisation hook or by a quarantine rule."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "No such resource."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Version conflict, state conflict or failed workflow guard."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "The request is well formed but cannot be applied."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Detect Keys",
        "tags": [
          "reference"
        ]
      }
    },
...
    "/records/{record_id}/conformance": {
      "get": {
        "description": "Conformance of the record's current values against its scope's effective schema.",
        "operationId": "get_conformance",
        "parameters": [
          {
            "in": "path",
            "name": "record_id",
            "required": true,
            "schema": {
              "title": "Record Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConformanceReport"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Missing or unknown bearer token."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Refused by the authorisation hook or by a quarantine rule."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "No such resource."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Version conflict, state conflict or failed workflow guard."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "The request is well formed but cannot be applied."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Conformance",
        "tags": [
          "reference"
        ]
      }
    },
...
    "/records/{record_id}/workflow": {
      "get": {
        "description": "State, state-entered time and every transition with its guard results.",
        "operationId": "get_workflow_status",
        "parameters": [
          {
            "in": "path",
            "name": "record_id",
            "required": true,
            "schema": {
              "title": "Record Id",
              "type": "string"
            }
          },
          {
            "description": "Roles for the role guards.",
            "in": "query",
            "name": "role",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Roles for the role guards.",
              "title": "Role"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowStatus"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Missing or unknown bearer token."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Refused by the authorisation hook or by a quarantine rule."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "No such resource."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Version conflict, state conflict or failed workflow guard."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "The request is well formed but cannot be applied."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Workflow Status",
        "tags": [
          "reference"
        ]
      }
    },
    "/relations": {
      "get": {
        "description": "The relation vocabulary in display order.",
        "operationId": "list_relations",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RelationOut"
                  },
                  "title": "Response List Relations",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Missing or unknown bearer token."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Refused by the authorisation hook or by a quarantine rule."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "No such resource."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Version conflict, state conflict or failed workflow guard."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "The request is well formed but cannot be applied."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Relations",
        "tags": [
          "reference"
        ]
      }
    },
    "/relations/default": {
      "get": {
        "description": "The relation to pre-select when linking a record of `from_type` to one of `to_type`.",
        "operationId": "get_default_relation",
        "parameters": [
          {
            "in": "query",
            "name": "from_type",
            "required": true,
            "schema": {
              "title": "From Type",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "to_type",
            "required": true,
            "schema": {
              "title": "To Type",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DefaultRelationOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Missing or unknown bearer token."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Refused by the authorisation hook or by a quarantine rule."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "No such resource."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Version conflict, state conflict or failed workflow guard."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "The request is well formed but cannot be applied."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Default Relation",
        "tags": [
          "reference"
        ]
      }
    },
    "/schema/forms": {
      "get": {
        "description": "Form and grid metadata for a record type under the scope's effective schema.",
        "operationId": "get_form_metadata",
        "parameters": [
          {
            "in": "query",
            "name": "scope",
            "required": true,
            "schema": {
              "title": "Scope",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "record_type",
            "required": true,
            "schema": {
              "title": "Record Type",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FormMetadata"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Missing or unknown bearer token."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Refused by the authorisation hook or by a quarantine rule."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "No such resource."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Version conflict, state conflict or failed workflow guard."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "The request is well formed but cannot be applied."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Form Metadata",
        "tags": [
          "reference"
        ]
      }
    },
...
  }
}